+++ tests/test_io_xarray.py
@@ -15,14 +15,14 @@ from rio_tiler.io import XarrayReader
 
 def test_xarray_reader():
     """test XarrayReader."""
-    arr = numpy.arange(0.0, 33 * 35).reshape(1, 33, 35)
+    arr = numpy.arange(0.0, 33 * 35 * 2).reshape(2, 33, 35)
     data = xarray.DataArray(
         arr,
         dims=("time", "y", "x"),
         coords={
             "x": numpy.arange(-170, 180, 10),
             "y": numpy.arange(-80, 85, 5),
-            "time": [datetime(2022, 1, 1)],
+            "time": [datetime(2022, 1, 1), datetime(2022, 1, 2)],
         },
     )
     data.attrs.update({"valid_min": arr.min(), "valid_max": arr.max()})
@@ -34,23 +34,50 @@ def test_xarray_reader():
         assert info.bounds == dst.bounds
         crs = info.crs
         assert rioCRS.from_user_input(crs) == dst.crs
-        assert info.band_metadata == [("b1", {})]
-        assert info.band_descriptions == [("b1", "2022-01-01T00:00:00.000000000")]
+        assert info.band_metadata == [("b1", {}), ("b2", {})]
+        assert info.band_descriptions == [
+            ("b1", "2022-01-01T00:00:00.000000000"),
+            ("b2", "2022-01-02T00:00:00.000000000"),
+        ]
         assert info.height == 33
         assert info.width == 35
-        assert info.count == 1
+        assert info.count == 2
         assert info.attrs
 
         stats = dst.statistics()
-        assert stats["2022-01-01T00:00:00.000000000"]
+        assert list(stats) == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
         assert stats["2022-01-01T00:00:00.000000000"].min == 0.0
 
+        stats = dst.statistics(indexes=1)
+        assert list(stats) == ["2022-01-01T00:00:00.000000000"]
+
+        stats = dst.statistics(indexes=2)
+        assert list(stats) == ["2022-01-02T00:00:00.000000000"]
+
+        stats = dst.statistics(indexes=(1, 2))
+        assert list(stats) == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
+
+        with pytest.raises(AssertionError):
+            stats = dst.statistics(indexes=(0,))
+
+        with pytest.raises(AssertionError):
+            stats = dst.statistics(indexes=0)
+
         img = dst.tile(0, 0, 0)
-        assert img.count == 1
+        assert img.count == 2
         assert img.width == 256
         assert img.height == 256
-        assert img.band_names == ["2022-01-01T00:00:00.000000000"]
-        assert img.dataset_statistics == ((arr.min(), arr.max()),)
+        assert img.band_names == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
+        assert img.dataset_statistics == ((arr.min(), arr.max()), (arr.min(), arr.max()))
 
         # Tests for auto_expand
         # Test that a high-zoom tile will error with auto_expand=False
@@ -64,59 +91,86 @@ def test_xarray_reader():
         assert "At least one of the clipped raster x,y coordinates" in str(error.value)
 
         # Test that a high-zoom tile will succeed with auto_expand=True (and that is the default)
-        img = dst.tile(tile.x, tile.y, zoom)
+        img = dst.tile(tile.x, tile.y, zoom, indexes=1)
         assert img.count == 1
         assert img.width == 256
         assert img.height == 256
         assert img.bounds == bounds
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
-        img = dst.part((-160, -80, 160, 80))
+        img = dst.part((-160, -80, 160, 80), indexes=1)
         assert img.crs == "epsg:4326"
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 33, 33)
 
-        img = dst.part((-160, -80, 160, 80), dst_crs="epsg:3857")
+        img = dst.part((-160, -80, 160, 80), dst_crs="epsg:3857", indexes=1)
         assert img.crs == "epsg:3857"
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 32, 34)
 
-        img = dst.part((-160, -80, 160, 80), max_size=15)
+        img = dst.part((-160, -80, 160, 80), max_size=15, indexes=1)
         assert img.array.shape == (1, 15, 15)
 
-        img = dst.part((-160, -80, 160, 80), width=40, height=35)
+        img = dst.part((-160, -80, 160, 80), width=40, height=35, indexes=1)
         assert img.array.shape == (1, 35, 40)
 
-        img = dst.part((-160, -80, 160, 80), max_size=15, resampling_method="bilinear")
+        img = dst.part(
+            (-160, -80, 160, 80), max_size=15, resampling_method="bilinear", indexes=1
+        )
         assert img.array.shape == (1, 15, 15)
 
         img = dst.preview()
+        assert img.crs == "epsg:4326"
+        assert img.count == 2
+        assert img.band_names == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
+        assert img.array.shape == (2, 33, 35)
+
+        img = dst.preview(indexes=1)
         assert img.crs == "epsg:4326"
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 33, 35)
 
-        img = dst.preview(dst_crs="epsg:3857")
+        img = dst.preview(dst_crs="epsg:3857", indexes=1)
         assert img.crs == "epsg:3857"
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 32, 36)
 
-        img = dst.preview(max_size=None)
+        img = dst.preview(max_size=None, indexes=1)
         assert img.array.shape == (1, 33, 35)
 
-        img = dst.preview(max_size=15)
+        img = dst.preview(max_size=15, indexes=1)
         assert img.array.shape == (1, 15, 15)
 
-        img = dst.preview(max_size=15, resampling_method="bilinear")
+        img = dst.preview(max_size=15, resampling_method="bilinear", indexes=1)
         assert img.array.shape == (1, 15, 15)
 
-        img = dst.preview(height=25, width=25, max_size=None)
+        img = dst.preview(height=25, width=25, max_size=None, indexes=1)
         assert img.array.shape == (1, 25, 25)
 
         pt = dst.point(0, 0)
+        assert pt.count == 2
+        assert pt.band_names == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
+        assert pt.coordinates
+        xys = [[0, 2.499], [0, 2.501], [-4.999, 0], [-5.001, 0], [-170, 80]]
+        for xy in xys:
+            x = xy[0]
+            y = xy[1]
+            pt = dst.point(x, y)
+            numpy.testing.assert_array_equal(
+                pt.data, data.sel(x=x, y=y, method="nearest").to_numpy()
+            )
+
+        pt = dst.point(0, 0, indexes=1)
         assert pt.count == 1
         assert pt.band_names == ["2022-01-01T00:00:00.000000000"]
         assert pt.coordinates
@@ -125,7 +179,9 @@ def test_xarray_reader():
             x = xy[0]
             y = xy[1]
             pt = dst.point(x, y)
-            assert pt.data[0] == data.sel(x=x, y=y, method="nearest")
+            assert pt.data[0] == data.sel(
+                time="2022-01-01T00:00:00.000000000", x=x, y=y, method="nearest"
+            )
 
         feat = {
             "type": "Feature",
@@ -148,21 +204,48 @@ def test_xarray_reader():
             },
         }
         img = dst.feature(feat)
+        assert img.count == 2
+        assert img.band_names == [
+            "2022-01-01T00:00:00.000000000",
+            "2022-01-02T00:00:00.000000000",
+        ]
+        assert img.array.shape == (2, 25, 32)
+
+        img = dst.feature(feat, indexes=1)
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 25, 32)
 
-        img = dst.feature(feat, dst_crs="epsg:3857")
+        img = dst.feature(feat, dst_crs="epsg:3857", indexes=1)
         assert img.count == 1
         assert img.band_names == ["2022-01-01T00:00:00.000000000"]
         assert img.crs == "epsg:3857"
         assert img.array.shape == (1, 20, 35)
 
-        img = dst.feature(feat, max_size=15)
+        img = dst.feature(feat, max_size=15, indexes=1)
         assert img.array.shape == (1, 12, 15)
 
-        img = dst.feature(feat, width=50, height=45)
+        img = dst.feature(feat, width=50, height=45, indexes=1)
         assert img.array.shape == (1, 45, 50)
+
+    # Select the first value
+    da = data[0]
+    assert da.ndim == 2
+    with XarrayReader(da) as dst:
+        assert dst.band_names == ["2022-01-01T00:00:00.000000000"]
+        info = dst.info()
+        assert info.band_descriptions == [("b1", "2022-01-01T00:00:00.000000000")]
+
+        stats = dst.statistics()
+        assert stats["2022-01-01T00:00:00.000000000"]
+        assert stats["2022-01-01T00:00:00.000000000"].min == 0.0
+
+        img = dst.tile(0, 0, 0)
+        assert img.count == 1
+        assert img.width == 256
+        assert img.height == 256
+        assert img.band_names == ["2022-01-01T00:00:00.000000000"]
+        assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
     arr = numpy.zeros((1, 1000, 2000))
     data = xarray.DataArray(
@@ -506,33 +589,45 @@ def test_xarray_reader_no_dims():
         crs = info.crs
         assert rioCRS.from_user_input(crs) == dst.crs
         assert info.band_metadata == [("b1", {})]
-        assert info.band_descriptions == [("b1", "value")]
+        assert info.band_descriptions == [("b1", "array")]
         assert info.height == 33
         assert info.width == 35
         assert info.count == 1
         assert info.attrs
 
         stats = dst.statistics()
-        assert stats["value"]
-        assert stats["value"].min == 0.0
+        assert stats["array"]
+        assert stats["array"].min == 0.0
+
+        stats = dst.statistics(indexes=1)
+        assert stats["array"]
+
+        stats = dst.statistics(indexes=(1,))
+        assert stats["array"]
+
+        with pytest.raises(ValueError):
+            stats = dst.statistics(indexes=2)
+
+        with pytest.raises(ValueError):
+            stats = dst.statistics(indexes=(1, 2))
 
         img = dst.tile(0, 0, 0)
         assert img.count == 1
         assert img.width == 256
         assert img.height == 256
-        assert img.band_names == ["value"]
+        assert img.band_names == ["array"]
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
         img = dst.part((-160, -80, 160, 80))
         assert img.count == 1
         assert img.width == 33
         assert img.height == 33
-        assert img.band_names == ["value"]
+        assert img.band_names == ["array"]
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
         pt = dst.point(0, 0)
         assert pt.count == 1
-        assert pt.band_names == ["value"]
+        assert pt.band_names == ["array"]
         assert pt.coordinates
         xys = [[0, 2.499], [0, 2.501], [-4.999, 0], [-5.001, 0], [-170, 80]]
         for xy in xys: